Rename mobile_ui directory to mobileui
