+++ mobile-ui/src/main/java/com/marshallaf/mobileui/model/UiProject.kt
@@ -1,0 +1,7 @@
+package com.marshallaf.mobileui.model
+
+data class UiProject(
+  val id: String, val name: String, val fullName: String, val starCount: String,
+  val dateCreated: String, val ownerName: String, val ownerAvatar: String,
+  val isBookmarked: Boolean
+)
